+++ client/Packages/com.beamable.server/CHANGELOG.md
@@ -7,11 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.17.0]
 ### Added
 - Ability to use Promises in ClientCallable methods
+- RemoteOnly MS visible in Miscroservice window
 - Container health checks are reported in deployment manifests
 - Local Mongo Storage Preview
 
 ### Fixed
 - Microservice clients can now deserialize json lists
+
+### Changed
+- Generated services no longer include the class name in the namespace
 
 ## [0.16.0]
 ### Added